Remove JobManager
