+++ packages/livebundle-github/test/GitHubAppServer.test.ts
@@ -3,11 +3,11 @@ import chaiHttp from "chai-http";
 import "mocha";
 import sinon from "sinon";
 import { GitHubAppServer } from "../src/GitHubAppServer";
-import { Job, JobManager, ServerConfig } from "../src/types";
+import { Job, JobRunner, ServerConfig } from "../src/types";
 
-class JobManagerNullImpl implements JobManager {
-  public add(job: Job): void {
-    // noop
+class JobRunnerNullImpl implements JobRunner {
+  public run(job: Job): Promise<void> {
+    return Promise.resolve();
   }
 }
 
@@ -21,29 +21,29 @@ describe("GitHubAppServer", () => {
 
   describe("get address", () => {
     it("should throw if the server is not started", () => {
-      const sut = new GitHubAppServer(serverConfig, new JobManagerNullImpl());
+      const sut = new GitHubAppServer(serverConfig, new JobRunnerNullImpl());
       expect(() => sut.address).to.throw();
     });
   });
 
   describe("get port", () => {
     it("should throw if the server is not started", () => {
-      const sut = new GitHubAppServer(serverConfig, new JobManagerNullImpl());
+      const sut = new GitHubAppServer(serverConfig, new JobRunnerNullImpl());
       expect(() => sut.port).to.throw();
     });
   });
 
   describe("functional tests", () => {
-    const jobManagerStub: sinon.SinonStubbedInstance<JobManagerNullImpl> = sinon.createStubInstance(
-      JobManagerNullImpl,
+    const jobManagerStub: sinon.SinonStubbedInstance<JobRunnerNullImpl> = sinon.createStubInstance(
+      JobRunnerNullImpl,
     );
-    let jobManagerAddStub: sinon.SinonStub<Job[], void>;
+    let jobRunnerRunStub: sinon.SinonStub<Job[], Promise<void>>;
 
     async function test(
       config: ServerConfig,
       func: (req: ChaiHttp.Agent, server: GitHubAppServer) => Promise<void>,
     ) {
-      const serv = new GitHubAppServer(config, new JobManagerNullImpl());
+      const serv = new GitHubAppServer(config, new JobRunnerNullImpl());
       try {
         await serv.start();
         await func(chai.request(`http://${serv.address}:${serv.port}`), serv);
@@ -53,11 +53,11 @@ describe("GitHubAppServer", () => {
     }
 
     beforeEach(() => {
-      jobManagerAddStub = jobManagerStub.add.returns();
+      jobRunnerRunStub = jobManagerStub.run.resolves();
     });
 
     afterEach(() => {
-      jobManagerAddStub.reset();
+      jobRunnerRunStub.reset();
     });
 
     describe("POST /", () => {
@@ -83,7 +83,7 @@ describe("GitHubAppServer", () => {
             repository: { owner: { login: "foo" }, name: "bar" },
             number: 456789,
           });
-          expect(jobManagerAddStub.calledOnce);
+          expect(jobRunnerRunStub.calledOnce);
         });
       });
     });
